Adds, verbatim the first three episodes of Unix Shell
Copies over verbatim the first three chapters of the [Unix Shell]() software carpentry.

This _not_ the final format. Revisions will be made to trim down, but before hacking at the material I wanted to get
_everything_ copied over and in subsequent revisions decide what material to keep.

Currently the example data is _not_ included as I figured we would customise this to the sample data that we set up the
course to use which currently is undecided.